A huge refactoring on the code.
Classes names changed, packages were created to separate the logic of the classes (models, views, data, activities).

The database storage was added and used on the activities.

The design of the parent part was almost all defined, just missing some dialog boxes and some additions to the UX.

Issue: The default image when adding a new question is a chair. We will change that later.

Missing features:
- Deleting items
- Dialog Boxes